refactor: 대규모 구조 리팩터링 🔁
- 더 이상 MVC 패턴을 사용하지 않음
- OutputView에서 버스트 상수를 Player가 아니라
Score에서 가져오도록 수정
- 패키지 구조 변경
