+++ src/main/java/BlackJackApp.java
@@ -1,13 +1,12 @@
-import controller.BlackJackController;
-
-import view.InputView;
-import view.OutputView;
+import object.game.BlackJackManager;
+import object.view.InputView;
+import object.view.OutputView;
 
 public class BlackJackApp {
     public static void main(String[] args) {
         final InputView inputView = new InputView();
         final OutputView outputView = new OutputView();
-        BlackJackController blackJackController = new BlackJackController(inputView, outputView);
-        blackJackController.run();
+        BlackJackManager blackJackManager = new BlackJackManager(inputView, outputView);
+        blackJackManager.run();
     }
 }
