--- src/main/java/domain/participant/ScoreType.java
@@ -1,8 +1,0 @@
-package domain.participant;
-
-public enum ScoreType {
-    BLACKJACK,
-    NORMAL,
-    BUST
-    ;
-}
