+++ src/main/java/dto/BattleResultResponse.java
@@ -1,7 +1,7 @@
 package dto;
 
-import domain.participant.BattleResult;
 import java.util.Map;
+import object.game.BattleResult;
 
 public record BattleResultResponse(String nickname, Map<BattleResult, Integer> battleResult) {
 }
