+++ src/main/java/object/game/ScoreType.java
@@ -1,0 +1,8 @@
+package object.game;
+
+public enum ScoreType {
+    BLACKJACK,
+    NORMAL,
+    BUST
+    ;
+}
