+++ src/main/java/object/participant/Participant.java
@@ -1,0 +1,12 @@
+package object.participant;
+
+import java.util.Map;
+import object.game.BattleResult;
+
+public interface Participant {
+    boolean ableToDraw(final int score);
+    boolean areYouDealer();
+    String getNickname();
+    void addGameRecord(BattleResult result);
+    Map<BattleResult, Integer> getGameRecord();
+}
