--- src/main/java/view/InputView.java
@@ -1,19 +1,0 @@
-package view;
-
-import java.util.List;
-import java.util.Scanner;
-
-public class InputView {
-    private static final Scanner scanner = new Scanner(System.in);
-
-    public List<String> askPlayerNames() {
-        System.out.println("게임에 참여할 사람의 이름을 입력하세요.(쉼표 기준으로 분리)");
-        String input = scanner.nextLine();
-        return List.of(input.split(","));
-    }
-
-    public String askDrawOneMore(String nickname) {
-        System.out.printf("%s는 한장의 카드를 더 받겠습니까?(예는 y, 아니오는 n)%n", nickname);
-        return scanner.nextLine();
-    }
-}
