+++ src/test/java/object/CardNumberTest.java
@@ -1,0 +1,16 @@
+package object;
+
+import object.card.CardNumber;
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.EnumSource;
+
+public class CardNumberTest {
+
+    @ParameterizedTest
+    @EnumSource(CardNumber.class)
+    void 카드_번호_생성_확인(CardNumber cardNumber) {
+        //then
+        Assertions.assertThat(cardNumber).isInstanceOf(CardNumber.class);
+    }
+}
