+++ src/test/java/object/DealerTest.java
@@ -1,0 +1,17 @@
+package object;
+
+import object.participant.Dealer;
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+public class DealerTest {
+
+    @Test
+    void 딜러_생성_테스트() {
+        //given
+        Dealer dealer = Dealer.generate();
+
+        //when & then
+        Assertions.assertThat(dealer).isInstanceOf(Dealer.class);
+    }
+}
