+++ src/test/java/object/ScoreTypeTest.java
@@ -1,0 +1,14 @@
+package object;
+
+import object.game.ScoreType;
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.EnumSource;
+
+public class ScoreTypeTest {
+    @ParameterizedTest
+    @EnumSource(ScoreType.class)
+    void ScoreType_생성_테스트(ScoreType scoreType) {
+        Assertions.assertThat(scoreType).isInstanceOf(ScoreType.class);
+    }
+}
